backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
Pillow==10.1.0
pypdf==3.17.4
python-docx==1.1.0
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
langchain==0.1.0
langchain-core==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20
openai==1.3.7
tiktoken==0.5.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
lxml==4.9.3
requests==2.31.0
beautifulsoup4==4.12.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk==1.35.0